--- src/domain/RacingConfig.js
@@ -1,1 +1,0 @@
-export const CAR_MOVE_CONDITION = 4;
